move merge internalTokenServiceImpl to TokenServiceImpl
