+++ impl/src/main/java/org/eclipse/tractusx/wallet/stub/token/impl/TokenServiceImpl.java
@@ -22,6 +22,8 @@
 
 package org.eclipse.tractusx.wallet.stub.token.impl;
 
+import com.nimbusds.jose.crypto.ECDSAVerifier;
+import com.nimbusds.jose.crypto.bc.BouncyCastleProviderSingleton;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
 import lombok.RequiredArgsConstructor;
@@ -46,6 +48,8 @@ import org.springframework.stereotype.Service;
 
 import java.nio.charset.StandardCharsets;
 import java.security.KeyPair;
+import java.security.interfaces.ECPublicKey;
+import java.text.ParseException;
 import java.util.Base64;
 import java.util.Date;
 import java.util.UUID;
@@ -53,11 +57,9 @@ import java.util.UUID;
 @Service
 @Slf4j
 @RequiredArgsConstructor
-public class TokenServiceImpl implements TokenService {
+public class TokenServiceImpl implements TokenService, InternalTokenValidationService {
 
     private final KeyService keyService;
-
-    private final InternalTokenValidationService internalTokenValidationService;
 
     private final DidDocumentService didDocumentService;
 
@@ -67,7 +69,7 @@ public class TokenServiceImpl implements TokenService {
     @Override
     public JWTClaimsSet verifyTokenAndGetClaims(String token) {
         try {
-            if (internalTokenValidationService.verifyToken(token)) {
+            if (verifyToken(token)) {
                 return SignedJWT.parse(CommonUtils.cleanToken(token)).getJWTClaimsSet();
             } else {
                 throw new IllegalArgumentException("Invalid token -> " + token);
@@ -139,6 +141,26 @@ public class TokenServiceImpl implements TokenService {
             }
         } catch (MalformedCredentialsException e) {
             throw e;
+        } catch (Exception e) {
+            throw new InternalErrorException("Internal Error: " + e.getMessage());
+        }
+    }
+
+    @Override
+    public boolean verifyToken(String token) {
+        try {
+            SignedJWT signedJWT = SignedJWT.parse(CommonUtils.cleanToken(token));
+            String keyID = signedJWT.getHeader().getKeyID(); //this will be DID
+            String bpn = CommonUtils.getBpnFromDid(keyID);
+            KeyPair keyPair = keyService.getKeyPair(bpn);
+            ECPublicKey aPublic = (ECPublicKey) keyPair.getPublic();
+            ECDSAVerifier ecdsaVerifier = new ECDSAVerifier(aPublic);
+            ecdsaVerifier.getJCAContext().setProvider(BouncyCastleProviderSingleton.getInstance());
+            return signedJWT.verify(ecdsaVerifier);
+        } catch (InternalErrorException e) {
+            throw e;
+        } catch (ParseException e) {
+            throw new ParseStubException(e.getMessage());
         } catch (Exception e) {
             throw new InternalErrorException("Internal Error: " + e.getMessage());
         }
